Delete the file OverviewScreen.kt, the screen that it was previously defining is now in MainScreen.kt, and so the OverViewScreenTest.kt is renamed MainScreenTest.kt
